Bug 1118510: Use the build submission time as the job start time
The previous behaviour was to use the actual build start/end time for
figuring out how many machines were needed at any one moment. In the
case where there are many pending jobs, the start time is delayed
relative to the job submission time, and the algorithm doesn't take that
into account. In addition, we calculate the optimal number of machines
here assuming that we can serve every request with a machine, but in
manage_jacuzzis.py we can assign multiple builders to a single machine.

If we use the submit time as the start time, we then are trying to
optimize the amount of time the pool is full/idle beginning from when
the jobs are submitted until they've actually finished. This should
adapt better to cases where too many jobs end up waiting a long time.

The net result of this is that we'll need more slaves allocated to most
builders.